Add a limbo context
Raft is a wild animal. If many members dissappear at the same time the remaining
nodes could fail to elect a new leader and enter in a never ending loop where
they all remain in candidate state.

It it happens, the limbo context returned by Start() will be canceled
by Rafty signaling the user that the rafty instance is done for and
should be recreated.


Signed-off-by: Sylvain Rabot <sylvain@abstraction.fr>